+++ environment_tests/test-exports-cjs/src/entrypoints.js
@@ -6,7 +6,6 @@ const base_language = require("langchain/base_language");
 const tools = require("langchain/tools");
 const chains = require("langchain/chains");
 const chains_openai_functions = require("langchain/chains/openai_functions");
-const chains_violation_of_expectations = require("langchain/chains/violation_of_expectations");
 const embeddings_base = require("langchain/embeddings/base");
 const embeddings_cache_backed = require("langchain/embeddings/cache_backed");
 const embeddings_fake = require("langchain/embeddings/fake");
@@ -82,5 +81,6 @@ const experimental_babyagi = require("langchain/experimental/babyagi");
 const experimental_generative_agents = require("langchain/experimental/generative_agents");
 const experimental_plan_and_execute = require("langchain/experimental/plan_and_execute");
 const experimental_chat_models_bittensor = require("langchain/experimental/chat_models/bittensor");
+const experimental_chains_violation_of_expectations = require("langchain/experimental/chains/violation_of_expectations");
 const evaluation = require("langchain/evaluation");
 const runnables_remote = require("langchain/runnables/remote");
