+++ environment_tests/test-exports-esbuild/src/entrypoints.js
@@ -6,7 +6,6 @@ import * as base_language from "langchain/base_language";
 import * as tools from "langchain/tools";
 import * as chains from "langchain/chains";
 import * as chains_openai_functions from "langchain/chains/openai_functions";
-import * as chains_violation_of_expectations from "langchain/chains/violation_of_expectations";
 import * as embeddings_base from "langchain/embeddings/base";
 import * as embeddings_cache_backed from "langchain/embeddings/cache_backed";
 import * as embeddings_fake from "langchain/embeddings/fake";
@@ -82,5 +81,6 @@ import * as experimental_babyagi from "langchain/experimental/babyagi";
 import * as experimental_generative_agents from "langchain/experimental/generative_agents";
 import * as experimental_plan_and_execute from "langchain/experimental/plan_and_execute";
 import * as experimental_chat_models_bittensor from "langchain/experimental/chat_models/bittensor";
+import * as experimental_chains_violation_of_expectations from "langchain/experimental/chains/violation_of_expectations";
 import * as evaluation from "langchain/evaluation";
 import * as runnables_remote from "langchain/runnables/remote";
