+++ environment_tests/test-exports-vite/src/entrypoints.js
@@ -6,7 +6,6 @@ export * from "langchain/base_language";
 export * from "langchain/tools";
 export * from "langchain/chains";
 export * from "langchain/chains/openai_functions";
-export * from "langchain/chains/violation_of_expectations";
 export * from "langchain/embeddings/base";
 export * from "langchain/embeddings/cache_backed";
 export * from "langchain/embeddings/fake";
@@ -82,5 +81,6 @@ export * from "langchain/experimental/babyagi";
 export * from "langchain/experimental/generative_agents";
 export * from "langchain/experimental/plan_and_execute";
 export * from "langchain/experimental/chat_models/bittensor";
+export * from "langchain/experimental/chains/violation_of_expectations";
 export * from "langchain/evaluation";
 export * from "langchain/runnables/remote";
