+++ langchain/.gitignore
@@ -64,9 +64,6 @@ chains/sql_db.d.ts
 chains/graph_qa/cypher.cjs
 chains/graph_qa/cypher.js
 chains/graph_qa/cypher.d.ts
-chains/violation_of_expectations.cjs
-chains/violation_of_expectations.js
-chains/violation_of_expectations.d.ts
 embeddings/base.cjs
 embeddings/base.js
 embeddings/base.d.ts
@@ -691,6 +688,9 @@ experimental/llms/bittensor.d.ts
 experimental/hubs/makersuite/googlemakersuitehub.cjs
 experimental/hubs/makersuite/googlemakersuitehub.js
 experimental/hubs/makersuite/googlemakersuitehub.d.ts
+experimental/chains/violation_of_expectations.cjs
+experimental/chains/violation_of_expectations.js
+experimental/chains/violation_of_expectations.d.ts
 evaluation.cjs
 evaluation.js
 evaluation.d.ts
