+++ langchain/src/experimental/chains/tests/violation_of_expectations_chain.int.test.ts
@@ -1,7 +1,7 @@
-import { ChatOpenAI } from "../../chat_models/openai.js";
-import { OpenAIEmbeddings } from "../../embeddings/openai.js";
-import { AIMessage, HumanMessage } from "../../schema/index.js";
-import { HNSWLib } from "../../vectorstores/hnswlib.js";
+import { ChatOpenAI } from "../../../chat_models/openai.js";
+import { OpenAIEmbeddings } from "../../../embeddings/openai.js";
+import { AIMessage, HumanMessage } from "../../../schema/index.js";
+import { HNSWLib } from "../../../vectorstores/hnswlib.js";
 import { ViolationOfExpectationsChain } from "../violation_of_expectations/violation_of_expectations_chain.js";
 
 const dummyMessages = [
