+++ langchain/src/experimental/chains/violation_of_expectations/index.ts
@@ -1,0 +1,4 @@
+export {
+  type ViolationOfExpectationsChainInput,
+  ViolationOfExpectationsChain,
+} from "./violation_of_expectations_chain.js";
