+++ langchain/src/load/import_map.ts
@@ -7,7 +7,6 @@ export * as base_language from "../base_language/index.js";
 export * as tools from "../tools/index.js";
 export * as chains from "../chains/index.js";
 export * as chains__openai_functions from "../chains/openai_functions/index.js";
-export * as chains__violation_of_expectations from "../chains/violation_of_expectations/violation_of_expectations_chain.js";
 export * as embeddings__base from "../embeddings/base.js";
 export * as embeddings__cache_backed from "../embeddings/cache_backed.js";
 export * as embeddings__fake from "../embeddings/fake.js";
@@ -83,5 +82,6 @@ export * as experimental__babyagi from "../experimental/babyagi/index.js";
 export * as experimental__generative_agents from "../experimental/generative_agents/index.js";
 export * as experimental__plan_and_execute from "../experimental/plan_and_execute/index.js";
 export * as experimental__chat_models__bittensor from "../experimental/chat_models/bittensor.js";
+export * as experimental__chains__violation_of_expectations from "../experimental/chains/violation_of_expectations/index.js";
 export * as evaluation from "../evaluation/index.js";
 export * as runnables__remote from "../runnables/remote.js";
