+++ langchain/tsconfig.json
@@ -54,7 +54,6 @@
       "src/chains/query_constructor/ir.ts",
       "src/chains/sql_db/index.ts",
       "src/chains/graph_qa/cypher.ts",
-      "src/chains/violation_of_expectations/violation_of_expectations_chain.ts",
       "src/embeddings/base.ts",
       "src/embeddings/bedrock.ts",
       "src/embeddings/cache_backed.ts",
@@ -263,6 +262,7 @@
       "src/experimental/chat_models/bittensor.ts",
       "src/experimental/llms/bittensor.ts",
       "src/experimental/hubs/makersuite/googlemakersuitehub.ts",
+      "src/experimental/chains/violation_of_expectations/index.ts",
       "src/evaluation/index.ts",
       "src/runnables/remote.ts"
     ],
